engine-3-vat/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pdfplumber==0.11.0
pyyaml==6.0.1
google-generativeai==0.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
jinja2==3.1.4
itsdangerous==2.2.0
httpx==0.27.0
